requirements.txt
# Core data
pandas==2.2.2
pyarrow==15.0.2
pyspark==3.5.1

# dbt
dbt-core==1.8.3
dbt-postgres==1.8.3

# Statistics
scipy==1.13.1
statsmodels==0.14.2
pymc==5.15.0
arviz==0.18.0

# ML
xgboost==2.0.3
scikit-learn==1.5.0

# API
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.3

# Frontend
streamlit==1.35.0
plotly==5.22.0

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# AWS
boto3==1.34.120
sagemaker==2.220.0

# Testing
pytest==8.2.2
pytest-cov==5.0.0
pytest-asyncio==0.23.7
httpx==0.27.0

# Linting
ruff==0.4.8
black==24.4.2
